Returns path to an isolated fixture file
Fixture file modifications by some of the tests are causing them to sometimes fail depending on the order that the tests are run in.

This copies the entire fixtures directory tree to a temp location, and then returns a path to the newly created temp file version of the requested file.